ms-collecte-iot/pyproject.toml
[tool.poetry]
name = "ms-collecte-iot"
version = "0.1.0"
description = "Microservice Collecte IoT — UrbanHub"
authors = ["Géraldine François <g9francois@gmail.com>"]
readme = "readme_collecte_iot.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = "^0.29.0"
pydantic = ">=2.0"
kafka-python = "^2.0.2"
pymongo = "^4.6"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pytest-cov = "^5.0.0"
pytest-asyncio = "^0.23.0"
httpx = "^0.27.0"
flake8 = "^7.0.0"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --junitxml=reports/junit.xml --cov=app --cov-report=term-missing --cov-report=xml:reports/coverage.xml --cov-report=html:reports/htmlcov"

[tool.flake8]
max-line-length = 120
exclude = [".git", "__pycache__", ".venv"]
ignore = ["E203", "W503"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
